phase3: split memory/init functions into src/memory.cpp
Extracted InitialiseOutputStreams, Initialise, InitialiseABC,
ReadInputPointcloud, ReadInputInventory, AllocMem, FreeMem
from troll.cpp into src/memory.cpp (-1178 lines).
All forward declarations were already present in troll.hpp.
