+++ CMakeLists.txt
@@ -25,6 +25,7 @@ add_executable(${PROJECT_NAME}
     src/tree.cpp
     src/lookup_tables.cpp
     src/input.cpp
+    src/memory.cpp
     src/lookUpTables_cache.cpp
     src/params/param_registry.cpp
 )
